devide getGrowiInfo as a service
