+++ apps/app/src/server/service/growi-info/index.ts
@@ -1,0 +1,1 @@
+export * from './growi-info';
